Parser

+++ src/error.hpp
@@ -13,14 +13,14 @@ std::string get_error_type_string(const ERROR_TYPE error_type);
 struct Error
 {
     const ERROR_TYPE type;
-    const long line;
-    const long column;
+    const size_t line;
+    const size_t column;
     const std::string file_name;
     const std::string message;
 
     Error() : type(ERROR_TYPE::ERROR), line(0L), column(0L), file_name(""), message("") {}
 
-    Error(ERROR_TYPE type, const long line, const long column, const std::string &file, const std::string &message)
+    Error(ERROR_TYPE type, const size_t line, const size_t column, const std::string &file, const std::string &message)
     : type(type), line(line), column(column), file_name(file), message(message) {}
 };
